* barryvdh/laravel-debugbar * barryvdh/laravel-ide-helper * немного раскидана модели

+++ database/migrations/2020_07_11_201329_create_tbl_recipe.php
@@ -17,7 +17,7 @@ class CreateTblRecipe extends Migration
             $table->id();
             $table->string('title');
             $table->text('description');
-            $table->integer('priority')->nullable();
+            $table->tinyInteger('priority')->nullable();
             $table->timestampsTz();
             $table->softDeletesTz();
         });
